+++ index.js
@@ -3,6 +3,7 @@ require("dotenv").config();
 const Discord = require("discord.js");
 const config = require("./configs/config.json");
 const wokCommands = require("wokcommands");
+const { status: s, interval } = require("./configs/botActivity");
 
 const client = new Discord.Client({
   partials: ["MESSAGE", "CHANNEL", "REACTION"],
@@ -11,20 +12,14 @@ const client = new Discord.Client({
 client.on("ready", () => {
   client.prefix = config.prefix;
   client.embedColor = config.embedColor;
-  if (config.botActivity.type === "multiple") {
-    const randomStaus = () => {
-      let random = config.botActivity.multiple.presence;
-      let rStatus = Math.floor(Math.random() * random.length);
-      client.user.setActivity(random[rStatus], {
-        type: config.botActivity.multiple.activity,
-      });
-    };
-    setInterval(randomStaus, 30000);
-  } else if (config.botActivity.type === "single") {
-    client.user.setActivity(config.botActivity.single.presence, {
-      type: config.botActivity.single.activity,
+  const statusOptions = s(client);
+  const randomStatus = () => {
+    const status = Math.floor(Math.random() * statusOptions.length);
+    client.user.setActivity(statusOptions[status].text, {
+      type: statusOptions[status].type,
     });
-  }
+  };
+  setInterval(randomStatus, interval);
 
   // client.user.setStatus(config.botActivity.status);
   const messagesPath = "./messages.json";